Refactor 'partialRight' code
